Move service/store methods from autoindexing to uploads.
