+++ enterprise/internal/codeintel/autoindexing/shared/types.go
@@ -4,17 +4,6 @@ import (
 	"github.com/sourcegraph/sourcegraph/enterprise/internal/codeintel/shared/types"
 )
 
-type GetIndexesOptions struct {
-	RepositoryID  int
-	State         string
-	States        []string
-	Term          string
-	IndexerNames  []string
-	WithoutUpload bool
-	Limit         int
-	Offset        int
-}
-
 type SourcedCommits struct {
 	RepositoryID   int
 	RepositoryName string
@@ -32,22 +21,6 @@ type IndexesWithRepositoryNamespace struct {
 	Root    string
 	Indexer string
 	Indexes []types.Index
-}
-
-type DeleteIndexesOptions struct {
-	States        []string
-	IndexerNames  []string
-	Term          string
-	RepositoryID  int
-	WithoutUpload bool
-}
-
-type ReindexIndexesOptions struct {
-	States        []string
-	IndexerNames  []string
-	Term          string
-	RepositoryID  int
-	WithoutUpload bool
 }
 
 type RepositoryWithCount struct {
