+++ enterprise/internal/codeintel/uploads/service.go
@@ -229,3 +229,31 @@ func (s *Service) ReindexUploads(ctx context.Context, opts shared.ReindexUploads
 func (s *Service) ReindexUploadByID(ctx context.Context, id int) error {
 	return s.store.ReindexUploadByID(ctx, id)
 }
+
+func (s *Service) GetIndexes(ctx context.Context, opts shared.GetIndexesOptions) ([]types.Index, int, error) {
+	return s.store.GetIndexes(ctx, opts)
+}
+
+func (s *Service) GetIndexByID(ctx context.Context, id int) (types.Index, bool, error) {
+	return s.store.GetIndexByID(ctx, id)
+}
+
+func (s *Service) GetIndexesByIDs(ctx context.Context, ids ...int) ([]types.Index, error) {
+	return s.store.GetIndexesByIDs(ctx, ids...)
+}
+
+func (s *Service) DeleteIndexByID(ctx context.Context, id int) (bool, error) {
+	return s.store.DeleteIndexByID(ctx, id)
+}
+
+func (s *Service) DeleteIndexes(ctx context.Context, opts shared.DeleteIndexesOptions) error {
+	return s.store.DeleteIndexes(ctx, opts)
+}
+
+func (s *Service) ReindexIndexByID(ctx context.Context, id int) error {
+	return s.store.ReindexIndexByID(ctx, id)
+}
+
+func (s *Service) ReindexIndexes(ctx context.Context, opts shared.ReindexIndexesOptions) error {
+	return s.store.ReindexIndexes(ctx, opts)
+}
